backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Data Validation & Settings
pydantic==2.4.2
pydantic-settings==2.0.3

# Environment & Configuration
python-dotenv==1.0.0

# HTTP & API
httpx==0.25.2
requests==2.31.0
fastapi-cors==0.0.6

# File Operations (for streaming)
aiofiles==23.2.0

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# Video Processing
ffmpeg-python==0.2.1

# Payment Processing
stripe==8.10.0
razorpay==1.3.0

# Cloud Storage
boto3==1.28.85

# Rate Limiting
slowapi==0.1.9

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1